refactor: business.d.ts의 타입 분리 #272
* CursorPage 타입 적용
