+++ src/features/bookmarks/types/index.d.ts
@@ -1,0 +1,8 @@
+declare interface Bookmark {
+  animeId: number;
+  title: string;
+  thumbnail: string;
+  createdAt: string;
+  avgScore: number;
+  myScore: number;
+}
